Crate xen

Interface with Xen hypervisor

Re-exports

pub use xen_sys;

Modules

console

Console utilities

events

Interface to Xen event channels

grant_table

Interface for Xen Grant Tables

hypercall

Platform-agnostic hypercall interface

memory

Memory utility functions and types

platform

Platform-specific

scheduler

Virtual machine scheduler interface

sections

Helper functions for section addresses

time

Blocking time

trap

Trap handling

xenbus

XenBus

xenstore

XenStore interface

Macros

dbg

Prints and returns the value of a given expression for quick and dirty debugging

hypercall

Software trap from a domain to the hypervisor used to request privileged operations

print

Prints to the Xen console

println

Prints to the Xen console with newline and carriage return

Structs

Delay

Future for delaying asynchronous execution for the supplied Duration

Constants

DOMID_SELF

Domain ID of this domain

Statics

SHARED_INFO

Xen dynamic global state information

START_INFO

Xen static startup information

xen_features

?

Functions

init_info

Map shared info page, initialise start and shared info pointers