Struct xen::grant_table::operations::GrantHandle

pub struct GrantHandle { /* private fields */ }

Handle to track a mapping created via a grant reference

Fields are private so that a handle cannot be constructed other than by map_grant_entry.

Implementations

impl GrantHandle

pub unsafe fn new(
address: *const u8,
reference: u32,
domain: domid_t,
readonly: bool
) -> Result<Self, Error>

Maps a new grant entry

pub fn unmap(&self) -> Result<(), Error>

Unmaps the mapped grant reference