Enum xen::hypercall::Error

pub enum Error {

    PERM,
    NOENT,
    SRCH,
    INTR,
    IO,
    NXIO,
    TOOBIG,
    NOEXEC,
    BADF,
    CHILD,
    WOULDBLOCK,
    NOMEM,
    ACCES,
    FAULT,
    BUSY,
    EXIST,
    XDEV,
    NODEV,
    NOTDIR,
    ISDIR,
    INVAL,
    NFILE,
    MFILE,
    NOSPC,
    ROFS,
    MLINK,
    DOM,
    RANGE,
    DEADLOCK,
    NAMETOOLONG,
    NOLCK,
    NOSYS,
    NOTEMPTY,
    NODATA,
    TIME,
    BADMSG,
    OVERFLOW,
    ILSEQ,
    RESTART,
    NOTSOCK,
    MSGSIZE,
    OPNOTSUPP,
    ADDRINUSE,
    ADDRNOTAVAIL,
    NOBUFS,
    ISCONN,
    NOTCONN,
    TIMEDOUT,
    CONNREFUSED,
    Unknown(i64),

}

Hypercall Error

Variants

PERM

Operation not permitted

NOENT

No such file or directory

SRCH

No such process

INTR

Interrupted system call. Internal only, should never be exposed to the guest.

IO

I/O error

NXIO

No such device or address

TOOBIG

Arg list too long

NOEXEC

Exec format error

BADF

Bad file number

CHILD

No child processes

WOULDBLOCK

Operation would block

NOMEM

Out of memory

ACCES

Permission denied

FAULT

Bad address

BUSY

Device or resource busy

EXIST

File exists

XDEV

Cross-device link

NODEV

No such device

NOTDIR

Not a directory

ISDIR

Is a directory

INVAL

Invalid argument

NFILE

File table overflow

MFILE

Too many open files

NOSPC

No space left on device

ROFS

Read-only file system

MLINK

Too many links

DOM

Math argument out of domain of func

RANGE

Math result not representable

DEADLOCK

Resource deadlock would occur

NAMETOOLONG

File name too long

NOLCK

No record locks available

NOSYS

Function not implemented

NOTEMPTY

Directory not empty

NODATA

No data available

TIME

Timer expired

BADMSG

Not a data message

OVERFLOW

Value too large for defined data type

ILSEQ

Illegal byte sequence

RESTART

Interrupted system call should be restarted. Internal only, should never be exposed to the guest.

NOTSOCK

Socket operation on non-socket

MSGSIZE

Message too large.

OPNOTSUPP

Operation not supported on transport endpoint

ADDRINUSE

Address already in use

ADDRNOTAVAIL

Cannot assign requested address

NOBUFS

No buffer space available

ISCONN

Transport endpoint is already connected

NOTCONN

Transport endpoint is not connected

TIMEDOUT

Connection timed out

CONNREFUSED

Connection refused

Unknown(i64)

Unknown error