Module xen::memory

Memory utility functions and types

Consists of wrapper types representing different kinds of memory locations. The following diagram describes the conversions between them:

┌──────────────────┐           ┌───────────────┐
│MachineFrameNumber│◀─────────▶│PageFrameNumber│
└──────────────────┘           └───────────────┘
     ▲        ▲                        ▲
     │        │      ┌─────────┐       │
     │        └──────│PageEntry│───┐   │
     │               └─────────┘   │   │
     │                             ▼   ▼
     │                         ┌──────────────┐       ┌──────────────┐
     └────────────────────────▶│VirtualAddress│◀─────▶│MachineAddress│
                               └──────────────┘       └──────────────┘
                                       ▲                      ▲
                                       │                      │
                                       │                      │
                                       ▼                      │
                               ┌───────────────┐              │
                               │PhysicalAddress│◀─────────────┘
                               └───────────────┘

Modules

page_table

Xen Paravirtualized Page Table Interface

Structs

MachineAddress

Machine address

MachineFrameNumber

Number of a page in the machine’s address space

PageEntry

Page Entry

PageFrameNumber

Number for page frame

PhysicalAddress

Pseudo-Physical address

TLBFlushFlags

Flags for update_va_mapping

VirtualAddress

Virtual address

Functions

get_current_pages

Gets the current number of reserved pages for the current domain

get_max_machine_frame_number

Gets the maximum machine frame number of mapped RAM in this system

get_max_pages

Gets the maximum number of reserved pages for the current domain

hypervisor_mmu_update

Updates an entry in a page table

pfn_up

Gives a page frame number after rounding the given address to the next page frame boundary

update_va_mapping

Update virtual address mapping