Enum xen::xenbus::State

pub enum State {
    Unknown,
    Initialising,
    InitWait,
    Initialised,
    Connected,
    Closing,
    Closed,
    Reconfiguring,
    Reconfigured,
}

State of XenBus connection

Variants

Unknown

Initial state of the device on the bus, before either end has been connected

Initialising

Backend is in process of initializing itself

InitWait

Entered by the back end while it is waiting for information before completing initialization

Initialised

Backend is now ready for connection

Connected

Normal state of the bus

Closing

Device has become unavailable, backend is no longer doing useful work

Closed

Both halves of driver have disconnected

Reconfiguring

Device is being reconfigured

Reconfigured

Device reconfiguring has finished