Module xen::xenstore

XenStore interface

“The XenStore is a storage system shared between Xen guests. It is a simple hierarchical storage system, maintained by Domain 0 and accessed via a shared memory page and an event channel.” - The Definitive Guide to the Xen Hypervisor, Chapter 8

Functions

domain_id

Read the current domain’s ID

init

Initialize XenStore

ls

List contents of directory

read

Read a key’s value from the XenStore

write

Write a key-value pair to the XenStore